Fix the query for filtering by multiple tags
The main thing I learned here is to be specific about the columns
we want to select. We can do `select([cls.TABLE])` to tell sqlalchemy
to grab all the columns *and* to use straightforward column labels like
`builds.id` or `t1.name`, that are unique and easy to reference later.
And this makes our queries smaller so they're easier to debug.

If we don't specify the columns, then sqlalchemy picks colum labels that
make things hard to work with. Instead of `table.id` it will just use
`id`, which will create conflicts when selecting joined tables, or
selecting from subqueries.

Alternatively, you can use `.apply_labels()`, but this has other issues.
The column names end up being `builds_id` or `t1_name` and you have to
construct ugly strings to get what you want out of the table.